Adds Google Login functionality
